Grouped sleuth functions and renamed internal one.
